+++ lib/src/impl/in_code_configuration_processor.dart
@@ -18,9 +18,8 @@ class InCodeConfigurationProcessor extends ConfigurationProcessor {
    * Process each configuration and generate ObjectDefinition per item.
    */
   void _process(InCodeConfiguration config) {
-    config.objects.forEach((String id, _InstanceObject value) {
-      // Pass through each object in configuration
-      _InstanceObject instanceObject = value;
+    // Pass through each object in configuration
+    config.objects.forEach((String id, _InstanceObject instanceObject) {
       // Create instance of empty InCodeObjectDefinition
       InCodeObjectDefinition od = new InCodeObjectDefinition();
       // It keeps reference on instance function