+++ test/src/scope_test_group.dart
@@ -1,4 +1,4 @@
-//Copyright (C) 2012 Sergey Akopkokhyants. All Rights Reserved.
+//Copyright (C) 2012-2013 Sergey Akopkokhyants. All Rights Reserved.
 //Author: akserg
 
 part of lost_dart_test;