+++ test/src/share_classes.dart
@@ -1,20 +1,24 @@
-//Copyright (C) 2012 Sergey Akopkokhyants. All Rights Reserved.
+//Copyright (C) 2012-2013 Sergey Akopkokhyants. All Rights Reserved.
 //Author: akserg
 
 part of lost_dart_test;
 
-class Baz{
+class Baz {
   String name;
 
   Baz([this.name = '']);
 }
 
-class Bar{
+class Bar {
   Baz baz;
   Bar(this.baz);
 }
 
-class Foo{
+class Foo {
   Bar bar;
+  
+  String getBarBazName() {
+    return bar.baz.name;
+  }
 }